Consolidate ParsedInputAddress into model project
