+++ client/src/it/scala/grasshopper/client/census/CensusClientSpec.scala
@@ -1,6 +1,6 @@
 package grasshopper.client.census
 
-import grasshopper.client.census.model.ParsedInputAddress
+import grasshopper.model.census.ParsedInputAddress
 import org.scalatest.{FlatSpec, MustMatchers}
 
 import scala.concurrent.Await
@@ -23,7 +23,7 @@ class CensusClientSpec extends FlatSpec with MustMatchers {
 
 
   "A 'geocode' request" must "geocode an address string" in {
-    val parsedAddress = ParsedInputAddress(3146, "M St NW", "20007", "DC")
+    val parsedAddress = ParsedInputAddress("3146", "M St NW", "20007", "DC")
     val maybeAddress = Await.result(CensusClient.geocode(parsedAddress), timeout)
     maybeAddress match {
       case Right(result) =>
@@ -40,7 +40,7 @@ class CensusClientSpec extends FlatSpec with MustMatchers {
   }
 
   it should "respond with address with synonym in State definition" in {
-    val parsedAddress = ParsedInputAddress(456, "Central Ave", "11516", "New York")
+    val parsedAddress = ParsedInputAddress("456", "Central Ave", "11516", "New York")
     val maybeAddress = Await.result(CensusClient.geocode(parsedAddress), timeout)
     maybeAddress match {
       case Right(result) =>
